Update ocw-studio.yaml
We want to support a cross-site collection of resources, so that the
workflow that the OCW content authors have for the current production
site can be supported on ocw-studio as well.

After adding support for a cross-site relation in mitodl/ocw-studio#813
we have what we need to support that - we just need to make a change to
the ocw-www configuration in order to add a field which exercises the
cross_site option in order for it to work.

pr #99