add yaml linting and formatting setup
This adds a setup based on prettier and yamllint which will lint +
format our yaml files, giving us some assurance that the yaml checked in
here has valid syntax.

Note that this does not verify that the yaml is _valid for our
configuration_, since that would require validating it using the schema
which is locked away in ocw-studio at present.

for #112
pr #118