config for course list, collection
this does two things:

first it changes 'course collections' to be course lists.

then it adds configuration for a new 'course collection' content type,
not to be confused with the old one. This is basically a "collection of
lists," and will render w/ a description and then the course lists which
have been added to it.

The idea is that this will be a good way to represent some of the more
complicated collections that currently exist on production OCW, where
there are different sections and whatnot. Basically, course authors
would author each of those sections as a separate course collection,
with a description and a list of courses, and then add them to a course
collection gallery in the desired order.

closes #142
pr #140